Main clicker
I have gotten a main clicker function with upgrades going today. Still working on the endgame button that will take you to credits so you can reset the game from there. Also want to show the amount of clicks and upgrades on the end credit page, need to figure out th elogic to this.